start abstracting out categories list from home
